third_party/blink/web_tests/platform/linux/external/wpt/html/cross-origin-opener-policy/reporting/navigation-reporting/reporting-redirect-with-same-origin-allow-popups.https-expected.txt

This is a testharness.js-based test.
FAIL Same origin openee redirected to same-origin with same-origin-allow-popups assert_equals: expected (string) "coop" but got (undefined) undefined
FAIL Cross origin openee redirected to same-origin with same-origin-allow-popups assert_equals: expected (string) "coop" but got (undefined) undefined
Harness: the test ran to completion.
